Website/requirements.txt
flask
flask-table
Adafruit-DHT
mysql-connector
RPi.GPIO